backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.1
Pillow==10.2.0
psycopg2-binary==2.9.9
PyMySQL==1.1.0
django-anymail==10.2
beautifulsoup4==4.12.3
requests==2.31.0
selenium==4.17.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
tqdm==4.66.1
